These optimizations improve the readability and performance of the code.
1.
Object Destructuring: Instead of using Object.assign, we can use object destructuring to merge the ownActionTypes and actionTypes objects.

2. Use Map for pastActions: Instead of using an object to store pastActions, we can use a Map for better performance and easier cleanup.

Use has method for Map and default value for get: We can use the has method of Map to check if an action type exists in pastActions instead of checking for undefined. Additionally, we can provide a default value to the get function to simplify the code.